server/requirements.txt
openai==1.60.1
Flask==3.1.0
python-dotenv-1.0.1